+++ dom/src/main/java/react4j/dom/proptypes/html/AudioProps.java
@@ -5,6 +5,7 @@ import javax.annotation.Nullable;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
+import org.intellij.lang.annotations.MagicConstant;
 import react4j.dom.events.DragEventHandler;
 import react4j.dom.events.FocusEventHandler;
 import react4j.dom.events.KeyboardEventHandler;
@@ -81,9 +82,9 @@ public class AudioProps
 
   @JsOverlay
   @Nonnull
-  public final AudioProps preLoad( AudioPreload s )
+  public final AudioProps preLoad( @MagicConstant( valuesFromClass = AudioPreload.class ) final String value )
   {
-    setPreload( s.name() );
+    setPreload( value );
     return self();
   }