+++ dom/src/main/java/react4j/dom/proptypes/html/BtnProps.java
@@ -5,6 +5,7 @@ import javax.annotation.Nullable;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
+import org.intellij.lang.annotations.MagicConstant;
 import react4j.dom.events.DragEventHandler;
 import react4j.dom.events.FocusEventHandler;
 import react4j.dom.events.FormEventHandler;
@@ -115,9 +116,9 @@ public class BtnProps
 
   @JsOverlay
   @Nonnull
-  public final BtnProps type( ButtonType t )
+  public final BtnProps type( @MagicConstant( valuesFromClass = ButtonType.class ) final String type )
   {
-    setType( t.name() );
+    setType( type );
     return self();
   }