+++ dom/src/main/java/react4j/dom/proptypes/html/FormProps.java
@@ -5,6 +5,7 @@ import javax.annotation.Nullable;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
+import org.intellij.lang.annotations.MagicConstant;
 import react4j.dom.events.DragEventHandler;
 import react4j.dom.events.FocusEventHandler;
 import react4j.dom.events.FormEventHandler;
@@ -40,9 +41,9 @@ public class FormProps
 
   @JsOverlay
   @Nonnull
-  public final FormProps autoComplete( OnOff s )
+  public final FormProps autoComplete( @MagicConstant( valuesFromClass = OnOff.class ) final String value )
   {
-    setAutoComplete( s.name() );
+    setAutoComplete( value );
     return self();
   }
 
@@ -56,9 +57,9 @@ public class FormProps
 
   @JsOverlay
   @Nonnull
-  public final FormProps formMethod( FormMethod s )
+  public final FormProps formMethod( @MagicConstant( valuesFromClass = FormMethod.class ) final String method )
   {
-    setFormMethod( s.name() );
+    setFormMethod( method );
     return self();
   }