+++ dom/src/main/java/react4j/dom/proptypes/html/attributeTypes/AudioPreload.java
@@ -1,6 +1,29 @@
 package react4j.dom.proptypes.html.attributeTypes;
 
-public enum AudioPreload
+import javax.annotation.Nonnull;
+
+/**
+ * An enumerated value to provide a hint to the browser about what the author thinks will lead to the best user experience.
+ */
+public final class AudioPreload
 {
-  auto, metadata, none
+  /**
+   * Indicates that the audio should not be preloaded.
+   */
+  @Nonnull
+  public static final String none = "none";
+  /**
+   * Indicates that only audio metadata (e.g. length) is fetched.
+   */
+  @Nonnull
+  public static final String metadata = "metadata";
+  /**
+   * Indicates that the whole audio file can be downloaded, even if the user is not expected to use it.
+   */
+  @Nonnull
+  public static final String auto = "auto";
+
+  private AudioPreload()
+  {
+  }
 }